requirements.txt
requests==2.20.0
django==3.2.13
PyYAML==5.4
mysqlclient==2.0.1
WebOb==1.5.0
ecdsa==0.13.3
httplib2==0.19.0
google-api-python-client==1.12.8
python-openid==2.2.5
pyasn1==0.4.6
pytz==2018.4
requests-oauthlib==0.7.0
rsa==4.7
simplejson==3.8.1
six==1.13.0
uritemplate==3.0.0
django-dotenv==1.4.2
django-allauth==0.41.0
pexpect==4.0.1
jsonschema==2.6.0
oauth2client==3.0.0
pyjwt==1.6.1
django-finalware==1.0.0
cryptography==3.3.2
future==0.18.3
google-cloud-storage==2.7.0
django-session-security==2.6.5
django-anymail[mailgun]
django-axes==5.20.0
django-redis==5.0.0
xhtml2pdf==0.2.9


# If you are seeing an error about django-debug-toolbar not being installed,
# the problem is almost certainly that your developer .env wasn't found
# django-debug-toolbar should **only** be installed on local developer builds, so do
# not add it to this file.

# If you are seeing an error about setuptools, the error is likely in your
# VM's build. Do not correct it with changes to requirements.txt!
